// ebk_front/home/bible/app/lib/config/interface.ts
export interface Eglise {
  createdAt: Date;
  updatedAt: Date;
  deletedAt: null;
  id_eglise: number;
  nom_eglise: string;
  username_eglise: string;
  photo_eglise: string;
  couverture_eglise: string;
  sigle_eglise: string;
  adresse_eglise: string;
  ville_eglise: string;
  pays_eglise: string;
  nombrefidel_eglise: string;
  status_eglise: string;
  payement_eglise: boolean;
  programme?: Programme[];
  word_of_day_eglise: string;
  about_eglise: string;
  localisation_eglise: string[];
}
// ...
export interface OccurenceSondage {
  createdAt: Date;
  updatedAt: Date;
  deletedAt: null;
  id: number;
  occurrence: string;
}
export interface QuestionnairesSondage {
  createdAt: Date;
  updatedAt: Date;
  deletedAt: null;
  id: number;
  question: string;
  type: SondageQuestionTypeEnum;
  occurrences: OccurenceSondage[];
  responses: [];
}
// ...
export interface ItemSondageQst {
  createdAt: Date;
  updatedAt: Date;
  deletedAt: null;
  id: number;
  title: string;
  objectif: string;
  public: boolean;
  eglise: Eglise;
  totalQuestion: number;
  totalAnswered: number;
}
export interface ItemSondageQstDetail extends ItemSondageQst {
  questions: QuestionnairesSondage[];
}